Only compress data at rest
We don't gain much by storing compressed cache values in memory, in fact
we end up allocating a lot of slices to decompress things.

This change introduces a buffer pool to help reduce GC churn. This is a
sufficient fix, but we also push compression down to only the PG layer.
This will make #28 easier to reason about.