Pull a few more fetches out of the denormalization step (#192)
This should ensure everything has already been fetched before we
denormalize data. This is important because the denorm step is
serialized (we don't want to add a work to an author before adding an
edition to the work). In cases where those fetches are cache misses, it
will enqueue redundant denorm steps but these will no-op. This could be
optimized but in practice this change is sufficient to keep the denorm
queue running smoothly.

The in-memory cache size is also reduced to 50% of available memory. 75%
combined with a large backlog of things to refresh made it too easy to
thrash GC.