Use legacy API for author lookups (#193)
The slowest part of the data loading pipeline is due to the step where
we resolve a legacy author/work IDs to their modern KCAs.

Currently, for works we're able to issue a HEAD request which redirects
us to the "best" book, which we can then fetch with gql to resolve
author/work KCA. For authors we need to load their works page, scrape
out book IDs, and then load those.

This works well enough but it could definitely be more efficient.
Fortunately the legacy API includes endpoints to do this more natively.
Unfortunately keys for the legacy API are now longer available...

However! The original R—— devs, in their infinite wisdom, were kind
enough to commit their API keys to source several years ago. I had
previously decided not to use these keys out of respect, but now that
the project is shut down I consider them fair game.

If/when the legacy API is retired we can revive this logic.