Resume author refreshes after reboot (#195)
Currently refreshing an author leaves it in an incomplete state until
we've fetched all of their works and denormalized them to the author. If
we shut down in the middle of that process the denorm step never
happens, so the author seems like they have no works.

With this change we store a bit for each author currently being
refreshed, and when we start up we check for those in-flight authors to
start refreshing them again.

It's still possible when things are backed up for the refresh and denorm
steps to happen very far apart, in which case users will see authors
missing works.

Fixes https://github.com/blampe/rreading-glasses/issues/90