Write to db even on cancellation (#197)
We sometimes skip writing to the DB in the case where our context has
been canceled e.g. due to request timeout. This is important data, since
these timeouts are typically due to the 1s throttle we have on upstream
requests.

Instead, if we have data to write we should always write it.