Fix an issue where authors could lose books during refresh (#218)
If an author took a long time to refresh (e.g. due to a backlog of other
work) we would end up returning the partial (one-book) author. Users
would see this as the author's works disappearing.

This PR fixes that by storing the author's last-known state along with
the fact that the author is being refreshed. We return that last-known
state for the entire duration of the refresh.

Fixes https://github.com/blampe/rreading-glasses/issues/173.