Add Cloudflare cache busting (#181)
This adds a new (optional) Cloudflare caching layer. This layer is a
no-op except on cache writes, when we queue a request to CF to bust the
corresponding URL. My reading of the [API
limits](https://developers.cloudflare.com/cache/how-to/purge-cache/#single-file-purge-limits)
is we are allowed 800 URL busts per second, which should be more than
enough.

I also updated the handler to return the current TTL instead of a fixed
one, so CF will stay close to our own refresh schedule. This will allow
me to turn on "cache everything" behavior while still keeping CF's cache
fresh.

I noticed the shutdown logic was leaving zombies so I disabled that and
can follow up in #90.