Fix stale data return (#222)
Expired data was still being returned as `data, false` which was causing
refresh logic to perform cold lookups instead of re-using KCAs. (Cold
lookups are extremely slow.)

This fixes the PG cache so the bool param is always true whenever data
was found. The ttl should be checked separately for expired-ness.