Ignore editions not attributed to the work's author
This should help cut back on a lot of unnecessary fetches for large
authors in particular.

It's not uncommon for an edition to belong to a work but not share the
work's author (according to the "best" edition). This most commonly
shows up in translations entirely attributed to the translator.

We should be able to safely ignore these. They currently just cause us
to load all of the translator's other editions, which isn't helpful.