Prefer FIFO order when recovering in-flight author refreshes (#233)
Sort recovered author refreshes so oldest are processed first.

I could do this more efficiently as part of the query, but I don't have
a compound index on (key, expires) and I'd rather not deal with a
migration if it can be avoided. I also don't think the overhead is worth
it for such a minor use case.