Buffer denorms in memory (#235)
We serialize all denormalization steps through a channel to prevent a
work attaching to an author before an edition attaches to the work.

Currently we spawn goroutines which wait until their turn to add to this
channel, but when the backlog is very large this can create hundreds of
thousands of idle goroutines. At that point performance begins to
degrade and CPU usage spikes due to most of our time being spent finding
goroutines to wake up.

This change immediately accepts writes to the channel but then buffers
them in memory before performing the denormalization. This is cheap in
terms of memory and it keeps CPU consistently in the ~30% with a full
backlog of work.

The buffer itself is an array of edges in the same order we pulled them
off the channel, along with indices into that array so we can de-dupe
operations on the same author/work. So as a side effect we also get much
smarter edge grouping behavior, which is helpful for reducing
deserializations.

The in-memory buffer isn't persisted across reboots, but this hasn't
been noticed by users and refreshing an author is a reasonable
workaround for now.