Improve series and primary author attribution (#281)
* Fix HC series to no longer include duplicated books in the same
position. Instead prefer the most popular book for that position.
* Fix an issue where series weren't getting cached or invalidated.
* Attach the complete series output to authors, instead of what we
derive from the author's works. This is more obviously correct, and the
client seems to correctly filter out works not belonging to the author.
* Authorship is now determined by the first contribution with a null
contribution type. This is less wrong that always taking the first
contribution, since that can be e.g. Illustrator. It would still be
preferable to support multiple authors in the client.