Simplify buffering (#303)
This simplifies our buffering logic and generalizes it to author
refreshes as well as the existing denormalization stuff. As a result I
haven't seen CPU go above 40% of one core while backlogged.

As a result this also cleans up some of the metrics awkwardness, since
we can now simply report the size of the buffer instead of trying to
compute as things are added and removed.

Also move all the controller background jobs to Run.